src/research-view/Touchless-Kiosk-Partner-Dashboard/requirements.txt
pillow==11.0.0
qrcode==8.0
leap==2021.1
macmouse==0.7.4